requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
deepface==0.0.93
filelock==3.16.1
fire==0.7.0
Flask==2.3.3
Flask-Cors==5.0.0
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.54.1
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
jax==0.4.38
jaxlib==0.4.38
Jinja2==3.1.4
joblib==1.4.2
keras==3.7.0
keyboard==0.13.5
kiwisolver==1.4.7
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==0.29.0
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.20
ml-dtypes==0.4.1
MouseInfo==0.1.3
mpmath==1.3.0
mtcnn==1.0.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
namex==0.0.8
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
protobuf==4.25.5
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyparsing==3.1.4
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytweening==1.2.0
pytz==2024.2
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
scipy==1.14.1
selenium==4.27.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.39
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
